JPA relationshsips

+++ src/main/java/com/example/week3/Entities/Entity.java
@@ -4,5 +4,5 @@ public interface Entity {
     public String getName();
     public void setName(String name);
     public int getId();
-    public void setId(int id);
+    public void setId(long id);
 }
